Add inspect signals.
Basic processor state inspection by wiring signals to registers. This basically adds async wires to the register file and introduces timing issues. The reason behind doing it anyway is keeping the educational aspect (it's easier to inspect registers then it is to use proper IO). Still, a proper IO system is planned.